move merkle tree logic away from shred definition
